CiCell: limit maximum height of stack trace <pre> container
Sometimes the stack traces can be many pages long for
each target variant (e.g. Construct) and in that case,
it doesn't make much sense to let them stretch to their
full height. This impairs usability, because it isn't
possible to quickly skim through all target variants
without having to scroll over the entire stack traces.

The most valuable information in stack traces is present
at the beginning anyway (error type and message + the most
recent stack frames), so I think it's reasonable to set
a max height to the <pre> element containing the trace,
showing an own scrollbar if necessary. This also makes it
much easier to compare errors across target variants
if the traces are long.