App: always show target variants in the same order
See https://github.com/kaitai-io/kaitai_struct/issues/823

Until now, the order in which variants are shown has
been somewhat "random". Not completely random, it was
just printed in the order in which the "ci.json" files
arrived. But the result was that for instance
the "javascript" variants were in a different order
every time: sometimes it was "node10, node12, node4, node8",
after reload some other one... This is not convenient to
work with, because you need to pay attention to the variant
labels, whose order is not consistent between reloads.