CiCell: darken the color of `passed` status without KST
The color of the `.passed` CSS class has been darkened, i.e. brought
closer to the `.passed-kst` class. To me, the color of `.passed` was
unnecessarily too bright, which made the stripes of `.passed` and
`.failed` colors in a cell with `mixed` status almost invisible. With
the new `.passed` color, you can see `.passed` / `.failed` stripes
relatively clearly.

Another reason why I think a darker `.passed` is better is because
`.passed` and `.passed-kst` both equally indicate that the test has
passed, there is no lesser `passed` status. It's OK to use a bit darker
green for tests generated from KST to suggest that we prefer to have
tests like this whenever possible. However, at this point, we're
generally using KST for all tests we can, and for each test we don't use
it, there's usually a reason for that (some unusual requirements on
asserts or the way the parser needs to be invoked).

Also, all colors have been shortened to the 3-valued `#RGB` syntax for
simplicity. I think having fewer options here actually makes the palette
easier to manage.