CiGrid: drop the "sort by column" functionality
To my knowledge, this feature has always been broken and never worked
properly. And I'm not even sure about its intended purpose. The only use
I can think of is to essentially group the tests by their status for the
particular column, so first would come all "failed", then all "passed"
etc. But I don't find that very useful, given that we already have a
text field for filtering targets and the "Only failures" checkbox, which
usually leaves only a few dozens of tests.

On top of that, I think the sorting feature in the current state does
more harm than good. I have activated it several times in the past by
clicking on some column header by accident, causing all tests to be
sorted in descending order (starting with "ZlibWithHeader78"), and
apparently it couldn't be deactivated without reloading the page.
Therefore, it's probably best to remove this broken feature for now.