Fix summarization of `unknown`+other status within one language
Until now, if a test had for example a `passed` status in one subtarget
of a certain language, but the same test was missing in the `ci.json`
for another subtarget of this language, the overall status of the
summarized cell would be `passed`. In other words, only subtargets that
had some status in their `ci.json` for a particular test were considered
for grouping. This is obviously wrong - the correct reporting is to set
the summarized status to `mixed` and indicate which subtargets have a
test status of `passed` and which `unknown`.

This commit fixes the merging algorithm to handle situations like this
correctly.